fix up paging
